==> HPC_Versions/python/4_logistic_regression.py <==
'''
Split data into a testing and training split.

We have three different problems we are investigating:
    1. Predicting face mask usage
    2. Predicting protective behaviour
    3. Predicting early time indicators

We have a single test/train split, and derive different response/features for each of the modelling goals.

Note to self: 
    Sensitivity = Recall, how many positives correctly predicted
    Precision: How many protected positives are correct
    
ToDo: Up-sampling on unbalanced data
    
Author:
    Jinjing Ye, Matt Ryan
    
Date created:
    17/04/2024
'''
# %% Packages
import pandas as pd
from sklearn.model_selection import KFold, cross_validate, StratifiedShuffleSplit
import numpy as np
from sklearn.linear_model import LogisticRegression
import pickle
from imblearn.over_sampling import RandomOverSampler
from sklearn.metrics import precision_score, recall_score, roc_auc_score, accuracy_score, f1_score

# %% Parameter set up

# initiate the model
logreg = LogisticRegression(max_iter=5000)
# kf = KFold(n_splits=5)
n_splits = 5
seed = 20240627
kf = StratifiedShuffleSplit(n_splits=n_splits,
                            test_size=1/n_splits,
                            random_state=seed)

model_fitting = "logistic_reg"
metric_list = ['precision', "recall", "roc_auc", "accuracy", "f1"]

# %%


def cross_validate_model(model_number, upsample=False):
    # Load data
    x = pd.read_csv(
        f"../data/X_train_{model_number}.csv", keep_default_na=False)
    y = pd.read_csv(f"../data/y_train_{model_number}.csv",
                    keep_default_na=False).values.ravel()

    # Cross validate model
    if upsample:
        cv_scores = {
            "fold": [],
            'test_precision': [],
            'test_recall': [],
            'test_roc_auc': [],
            'test_accuracy': [],
            'test_f1': []
        }

        splits = list(kf.split(x, y))

        for fold in range(len(splits)):
            cv_scores["fold"].append(fold)
            train_idx = splits[fold][0]
            val_idx = splits[fold][1]

            X_train = x.iloc[train_idx]
            y_train = y[train_idx]

            X_val = x.iloc[val_idx]
            y_val = y[val_idx]

            upsampler = RandomOverSampler()  # No seed set on purpose

            X_train_upsample, y_train_upsample = upsampler.fit_resample(
                X_train, y_train)

            clf = logreg.fit(X_train_upsample, y_train_upsample)

            preds = clf.predict(X_val)
            prop_preds = clf.predict_proba(X_val)

            cv_scores["test_precision"].append(
                precision_score(y_true=y_val, y_pred=preds))
            cv_scores["test_recall"].append(
                recall_score(y_true=y_val, y_pred=preds))
            cv_scores["test_roc_auc"].append(roc_auc_score(
                y_true=y_val, y_score=prop_preds[:, 1]))
            cv_scores["test_accuracy"].append(
                accuracy_score(y_true=y_val, y_pred=preds))
            cv_scores["test_f1"].append(f1_score(y_true=y_val, y_pred=preds))

        cv_scores["test_precision"] = np.array(cv_scores["test_precision"])
        cv_scores["test_recall"] = np.array(cv_scores["test_recall"])
        cv_scores["test_roc_auc"] = np.array(cv_scores["test_roc_auc"])
        cv_scores["test_accuracy"] = np.array(cv_scores["test_accuracy"])
        cv_scores["test_f1"] = np.array(cv_scores["test_f1"])
    else:
        cv_scores = cross_validate(logreg, x, y, cv=kf, scoring=metric_list)

    # Print the accuracy scores for each fold

    print(model_number)
    # print("Cross-validation scores:", cv_scores)

    print("Mean recall: ", cv_scores["test_recall"].mean().round(3))
    print("Mean roc: ", cv_scores["test_roc_auc"].mean().round(3))
    print("Mean accuracy: ", cv_scores["test_accuracy"].mean().round(3))

    # Save results

    with open(f"../results/{model_number}_{model_fitting}.pkl", "wb") as f:
        pickle.dump(cv_scores, f)

# %% Model 1:


model_number = "model_1"

cross_validate_model(model_number)

# %% Model 1a:

model_number = "model_1a"

cross_validate_model(model_number, upsample=True)

# %% Model 1b:

model_number = "model_1b"

cross_validate_model(model_number, upsample=True)

# %% Model 2:

model_number = "model_2"

cross_validate_model(model_number)

# %% Model 2a:

model_number = "model_2a"

cross_validate_model(model_number, upsample=True)

# %% Model 2b:

model_number = "model_2b"

cross_validate_model(model_number, upsample=True)
